feat(dj): Introduce getMeetingSession replacement, meetingSessionSummary
This PR deprecates getMeetingSession() and the events and types associated
with it. In their place is the new meetingSessionSummary() function and
meeting-session-summary-updated event. These are more properly named to
fit alongside the likes of meetingSessionState() as well as cache the
session summary stats so that querying them is not an asynchronous
operation. This should vastly improve today's behavior.